+++ dart/test/sentry_client_test.dart
@@ -6,6 +6,7 @@ import 'package:collection/collection.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/client_reports/discard_reason.dart';
 import 'package:sentry/src/client_reports/noop_client_report_recorder.dart';
+import 'package:sentry/src/platform/mock_platform.dart';
 import 'package:sentry/src/sentry_item_type.dart';
 import 'package:sentry/src/sentry_stack_trace_factory.dart';
 import 'package:sentry/src/sentry_tracer.dart';
@@ -19,8 +20,6 @@ import 'package:test/test.dart';
 import 'mocks.dart';
 import 'mocks/mock_client_report_recorder.dart';
 import 'mocks/mock_hub.dart';
-import 'mocks/mock_platform.dart';
-import 'mocks/mock_platform_checker.dart';
 import 'mocks/mock_transport.dart';
 import 'test_utils.dart';
 
@@ -1003,16 +1002,17 @@ void main() {
     });
   });
 
-  group('SentryClient: sets user & user ip', () {
+  group('SentryClient: user & user ip', () {
     late Fixture fixture;
 
     setUp(() {
       fixture = Fixture();
     });
 
-    test('event has no user', () async {
+    test('event has no user and sendDefaultPii = true', () async {
       final client = fixture.getSut(sendDefaultPii: true);
       var fakeEvent = SentryEvent();
+      expect(fakeEvent.user, isNull);
 
       await client.captureEvent(fakeEvent);
 
@@ -1021,12 +1021,27 @@ void main() {
 
       expect(fixture.transport.envelopes.length, 1);
       expect(capturedEvent.user, isNotNull);
-      expect(capturedEvent.user?.ipAddress, '{{auto}}');
+      expect(capturedEvent.user?.ipAddress, defaultIpAddress);
+    });
+
+    test('event has no user and sendDefaultPii = false', () async {
+      final client = fixture.getSut(sendDefaultPii: false);
+      var fakeEvent = SentryEvent();
+      expect(fakeEvent.user, isNull);
+
+      await client.captureEvent(fakeEvent);
+
+      final capturedEnvelope = fixture.transport.envelopes.first;
+      final capturedEvent = await eventFromEnvelope(capturedEnvelope);
+
+      expect(fixture.transport.envelopes.length, 1);
+      expect(capturedEvent.user, isNull);
     });
 
     test('event has a user with IP address', () async {
       final client = fixture.getSut(sendDefaultPii: true);
 
+      expect(fakeEvent.user?.ipAddress, isNotNull);
       await client.captureEvent(fakeEvent);
 
       final capturedEnvelope = fixture.transport.envelopes.first;
@@ -1040,10 +1055,12 @@ void main() {
       expect(capturedEvent.user?.email, fakeEvent.user!.email);
     });
 
-    test('event has a user without IP address', () async {
+    test('event has a user without IP address and sendDefaultPii = true',
+        () async {
       final client = fixture.getSut(sendDefaultPii: true);
 
       final event = fakeEvent.copyWith(user: fakeUser);
+      expect(event.user?.ipAddress, isNull);
 
       await client.captureEvent(event);
 
@@ -1052,7 +1069,26 @@ void main() {
 
       expect(fixture.transport.envelopes.length, 1);
       expect(capturedEvent.user, isNotNull);
-      expect(capturedEvent.user?.ipAddress, '{{auto}}');
+      expect(capturedEvent.user?.ipAddress, defaultIpAddress);
+      expect(capturedEvent.user?.id, fakeUser.id);
+      expect(capturedEvent.user?.email, fakeUser.email);
+    });
+
+    test('event has a user without IP address and sendDefaultPii = false',
+        () async {
+      final client = fixture.getSut(sendDefaultPii: false);
+
+      final event = fakeEvent.copyWith(user: fakeUser);
+      expect(event.user?.ipAddress, isNull);
+
+      await client.captureEvent(event);
+
+      final capturedEnvelope = fixture.transport.envelopes.first;
+      final capturedEvent = await eventFromEnvelope(capturedEnvelope);
+
+      expect(fixture.transport.envelopes.length, 1);
+      expect(capturedEvent.user, isNotNull);
+      expect(capturedEvent.user?.ipAddress, isNull);
       expect(capturedEvent.user?.id, fakeUser.id);
       expect(capturedEvent.user?.email, fakeUser.email);
     });
@@ -1869,9 +1905,7 @@ void main() {
     test(
         'Spotlight enabled should not set transport to SpotlightHttpTransport on iOS',
         () async {
-      fixture.options.platformChecker = MockPlatformChecker(
-        platform: MockPlatform.iOS(),
-      );
+      fixture.options.platform = MockPlatform.iOS();
       fixture.options.spotlight = Spotlight(enabled: true);
       fixture.getSut();
 
@@ -1881,9 +1915,7 @@ void main() {
     test(
         'Spotlight enabled should not set transport to SpotlightHttpTransport on macOS',
         () async {
-      fixture.options.platformChecker = MockPlatformChecker(
-        platform: MockPlatform.macOS(),
-      );
+      fixture.options.platform = MockPlatform.macOS();
       fixture.options.spotlight = Spotlight(enabled: true);
       fixture.getSut();
 
@@ -1893,9 +1925,7 @@ void main() {
     test(
         'Spotlight enabled should not set transport to SpotlightHttpTransport on Android',
         () async {
-      fixture.options.platformChecker = MockPlatformChecker(
-        platform: MockPlatform.android(),
-      );
+      fixture.options.platform = MockPlatform.android();
       fixture.options.spotlight = Spotlight(enabled: true);
       fixture.getSut();
 
@@ -1905,7 +1935,7 @@ void main() {
     test(
         'Spotlight enabled should set transport to SpotlightHttpTransport on Web',
         () async {
-      fixture.options.platformChecker = MockPlatformChecker(isWebValue: true);
+      fixture.options.platform = MockPlatform(isWeb: true);
       fixture.options.spotlight = Spotlight(enabled: true);
       fixture.getSut();
 
@@ -1915,8 +1945,7 @@ void main() {
     test(
         'Spotlight enabled should set transport to SpotlightHttpTransport on Linux',
         () async {
-      fixture.options.platformChecker =
-          MockPlatformChecker(platform: MockPlatform.linux());
+      fixture.options.platform = MockPlatform.linux();
       fixture.options.spotlight = Spotlight(enabled: true);
       fixture.getSut();
 
@@ -1926,8 +1955,7 @@ void main() {
     test(
         'Spotlight enabled should set transport to SpotlightHttpTransport on Windows',
         () async {
-      fixture.options.platformChecker =
-          MockPlatformChecker(platform: MockPlatform.windows());
+      fixture.options.platform = MockPlatform.windows();
       fixture.options.spotlight = Spotlight(enabled: true);
       fixture.getSut();
 
@@ -2261,8 +2289,7 @@ class Fixture {
   final recorder = MockClientReportRecorder();
   final transport = MockTransport();
 
-  final options =
-      defaultTestOptions(MockPlatformChecker(platform: MockPlatform.iOS()));
+  final options = defaultTestOptions()..platform = MockPlatform.iOS();
 
   late SentryTransactionContext _context;
   late SentryTracer tracer;