+++ flutter/lib/src/event_processor/flutter_enricher_event_processor.dart
@@ -19,8 +19,8 @@ class FlutterEnricherEventProcessor implements EventProcessor {
 
   final SentryFlutterOptions _options;
 
-  bool get _hasNativeIntegration => _checker.hasNativeIntegration;
-  PlatformChecker get _checker => _options.platformChecker;
+  bool get _hasNativeIntegration => _options.platform.supportsNativeIntegration;
+  RuntimeChecker get _checker => _options.runtimeChecker;
 
   // We can't use `WidgetsBinding` as a direct parameter
   // because it must be called inside the `runZoneGuarded`-Integration.
@@ -40,7 +40,7 @@ class FlutterEnricherEventProcessor implements EventProcessor {
     // information available than Flutter.
     // TODO: while we have a native integration with JS SDK, it's currently opt in and we dont gather contexts yet
     // so for web it's still better to rely on the information of Flutter.
-    final device = _hasNativeIntegration && !_checker.isWeb
+    final device = _hasNativeIntegration && !_options.platform.isWeb
         ? null
         : _getDevice(event.contexts.device);
 
@@ -138,8 +138,6 @@ class FlutterEnricherEventProcessor implements EventProcessor {
     // ignore: deprecated_member_use
     final hasRenderView = _widgetsBinding?.renderViewElement != null;
 
-    final renderer = _options.rendererWrapper.getRenderer()?.name;
-
     return <String, String>{
       'has_render_view': hasRenderView.toString(),
       if (tempDebugBrightnessOverride != null)
@@ -156,7 +154,6 @@ class FlutterEnricherEventProcessor implements EventProcessor {
       // Also always fails in tests.
       // See https://github.com/flutter/flutter/issues/83919
       // 'window_is_visible': _window.viewConfiguration.visible,
-      if (renderer != null) 'renderer': renderer,
       if (appFlavor != null) 'appFlavor': appFlavor!,
     };
   }
@@ -209,7 +206,7 @@ class FlutterEnricherEventProcessor implements EventProcessor {
     // See
     // - https://flutter.dev/docs/testing/build-modes
     // - https://github.com/flutter/flutter/wiki/Flutter%27s-modes
-    if (_checker.isWeb) {
+    if (_options.platform.isWeb) {
       if (_checker.isDebugMode()) {
         compiler = 'dartdevc';
       } else if (_checker.isReleaseMode() || _checker.isProfileMode()) {