+++ flutter/lib/src/web/sentry_js_binding.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/cupertino.dart';
 
 export 'noop_sentry_js_binding.dart'
-    if (dart.html) 'html_sentry_js_binding.dart'
     if (dart.library.js_interop) 'web_sentry_js_binding.dart';
 
 abstract class SentryJsBinding {