+++ flutter/test/mocks.dart
@@ -12,7 +12,7 @@ import 'package:sentry_flutter/src/frames_tracking/sentry_delayed_frames_tracker
 import 'package:sentry_flutter/src/native/sentry_native_binding.dart';
 import 'package:sentry_flutter/src/renderer/renderer.dart';
 import 'package:sentry_flutter/src/web/sentry_js_binding.dart';
-import 'package:platform/platform.dart';
+import 'package:sentry/src/platform/platform.dart';
 
 import 'mocks.mocks.dart';
 import 'no_such_method_provider.dart';
@@ -20,8 +20,10 @@ import 'no_such_method_provider.dart';
 const fakeDsn = 'https://abc@def.ingest.sentry.io/1234567';
 const fakeProguardUuid = '3457d982-65ef-576d-a6ad-65b5f30f49a5';
 
-SentryFlutterOptions defaultTestOptions([PlatformChecker? checker]) {
-  return SentryFlutterOptions(dsn: fakeDsn, checker: checker)
+SentryFlutterOptions defaultTestOptions(
+    {Platform? platform, RuntimeChecker? checker}) {
+  return SentryFlutterOptions(
+      dsn: fakeDsn, platform: platform, checker: checker)
     ..automatedTestMode = true;
 }
 
@@ -59,65 +61,29 @@ ISentrySpan startTransactionShim(
 ])
 void main() {}
 
-extension MockPlatform on FakePlatform {
-  static Platform android() {
-    return FakePlatform(operatingSystem: 'android');
-  }
-
-  static Platform iOS() {
-    return FakePlatform(operatingSystem: 'ios');
-  }
-
-  static Platform macOS() {
-    return FakePlatform(operatingSystem: 'macos');
-  }
-
-  static Platform linux() {
-    return FakePlatform(operatingSystem: 'linux');
-  }
-
-  static Platform windows() {
-    return FakePlatform(operatingSystem: 'windows');
-  }
-}
-
-class MockPlatformChecker with NoSuchMethodProvider implements PlatformChecker {
-  MockPlatformChecker({
-    this.buildMode = MockPlatformCheckerBuildMode.debug,
-    this.isWebValue = false,
-    this.hasNativeIntegration = false,
+class MockRuntimeChecker with NoSuchMethodProvider implements RuntimeChecker {
+  MockRuntimeChecker({
+    this.buildMode = MockRuntimeCheckerBuildMode.debug,
     this.isRoot = true,
-    Platform? mockPlatform,
-  }) : _mockPlatform = mockPlatform ?? FakePlatform(operatingSystem: '');
+  });
 
-  final MockPlatformCheckerBuildMode buildMode;
-  final bool isWebValue;
+  final MockRuntimeCheckerBuildMode buildMode;
   final bool isRoot;
-  final Platform _mockPlatform;
 
   @override
-  bool hasNativeIntegration = false;
+  bool isDebugMode() => buildMode == MockRuntimeCheckerBuildMode.debug;
 
   @override
-  bool isDebugMode() => buildMode == MockPlatformCheckerBuildMode.debug;
+  bool isProfileMode() => buildMode == MockRuntimeCheckerBuildMode.profile;
 
   @override
-  bool isProfileMode() => buildMode == MockPlatformCheckerBuildMode.profile;
-
-  @override
-  bool isReleaseMode() => buildMode == MockPlatformCheckerBuildMode.release;
+  bool isReleaseMode() => buildMode == MockRuntimeCheckerBuildMode.release;
 
   @override
   bool get isRootZone => isRoot;
-
-  @override
-  bool get isWeb => isWebValue;
-
-  @override
-  Platform get platform => _mockPlatform;
 }
 
-enum MockPlatformCheckerBuildMode { debug, profile, release }
+enum MockRuntimeCheckerBuildMode { debug, profile, release }
 
 // Does nothing or returns default values.
 // Useful for when a Hub needs to be passed but is not used.