+++ flutter/test/mocks.mocks.dart
@@ -189,9 +189,8 @@ class _FakeAccessibilityFeatures_25 extends _i1.SmartFake
       : super(parent, parentInvocation);
 }
 
-class _FakePipelineOwner_26 extends _i1.SmartFake
-    implements _i10.PipelineOwner {
-  _FakePipelineOwner_26(Object parent, Invocation parentInvocation)
+class _FakeRenderView_26 extends _i1.SmartFake implements _i10.RenderView {
+  _FakeRenderView_26(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 
   @override
@@ -199,28 +198,19 @@ class _FakePipelineOwner_26 extends _i1.SmartFake
       super.toString();
 }
 
-class _FakeRenderView_27 extends _i1.SmartFake implements _i10.RenderView {
-  _FakeRenderView_27(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-
-  @override
-  String toString({_i4.DiagnosticLevel? minLevel = _i4.DiagnosticLevel.info}) =>
-      super.toString();
-}
-
-class _FakeMouseTracker_28 extends _i1.SmartFake implements _i10.MouseTracker {
-  _FakeMouseTracker_28(Object parent, Invocation parentInvocation)
+class _FakeMouseTracker_27 extends _i1.SmartFake implements _i10.MouseTracker {
+  _FakeMouseTracker_27(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakePlatformMenuDelegate_29 extends _i1.SmartFake
+class _FakePlatformMenuDelegate_28 extends _i1.SmartFake
     implements _i9.PlatformMenuDelegate {
-  _FakePlatformMenuDelegate_29(Object parent, Invocation parentInvocation)
+  _FakePlatformMenuDelegate_28(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeFocusManager_30 extends _i1.SmartFake implements _i9.FocusManager {
-  _FakeFocusManager_30(Object parent, Invocation parentInvocation)
+class _FakeFocusManager_29 extends _i1.SmartFake implements _i9.FocusManager {
+  _FakeFocusManager_29(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 
   @override
@@ -228,52 +218,52 @@ class _FakeFocusManager_30 extends _i1.SmartFake implements _i9.FocusManager {
       super.toString();
 }
 
-class _FakeFuture_31<T1> extends _i1.SmartFake implements _i11.Future<T1> {
-  _FakeFuture_31(Object parent, Invocation parentInvocation)
+class _FakeFuture_30<T1> extends _i1.SmartFake implements _i11.Future<T1> {
+  _FakeFuture_30(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeCodec_32 extends _i1.SmartFake implements _i6.Codec {
-  _FakeCodec_32(Object parent, Invocation parentInvocation)
+class _FakeCodec_31 extends _i1.SmartFake implements _i6.Codec {
+  _FakeCodec_31(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeSemanticsHandle_33 extends _i1.SmartFake
+class _FakeSemanticsHandle_32 extends _i1.SmartFake
     implements _i12.SemanticsHandle {
-  _FakeSemanticsHandle_33(Object parent, Invocation parentInvocation)
+  _FakeSemanticsHandle_32(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeSemanticsUpdateBuilder_34 extends _i1.SmartFake
+class _FakeSemanticsUpdateBuilder_33 extends _i1.SmartFake
     implements _i6.SemanticsUpdateBuilder {
-  _FakeSemanticsUpdateBuilder_34(Object parent, Invocation parentInvocation)
+  _FakeSemanticsUpdateBuilder_33(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeViewConfiguration_35 extends _i1.SmartFake
+class _FakeViewConfiguration_34 extends _i1.SmartFake
     implements _i10.ViewConfiguration {
-  _FakeViewConfiguration_35(Object parent, Invocation parentInvocation)
+  _FakeViewConfiguration_34(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeSceneBuilder_36 extends _i1.SmartFake implements _i6.SceneBuilder {
-  _FakeSceneBuilder_36(Object parent, Invocation parentInvocation)
+class _FakeSceneBuilder_35 extends _i1.SmartFake implements _i6.SceneBuilder {
+  _FakeSceneBuilder_35(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakePictureRecorder_37 extends _i1.SmartFake
+class _FakePictureRecorder_36 extends _i1.SmartFake
     implements _i6.PictureRecorder {
-  _FakePictureRecorder_37(Object parent, Invocation parentInvocation)
+  _FakePictureRecorder_36(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeCanvas_38 extends _i1.SmartFake implements _i6.Canvas {
-  _FakeCanvas_38(Object parent, Invocation parentInvocation)
+class _FakeCanvas_37 extends _i1.SmartFake implements _i6.Canvas {
+  _FakeCanvas_37(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeWidget_39 extends _i1.SmartFake implements _i9.Widget {
-  _FakeWidget_39(Object parent, Invocation parentInvocation)
+class _FakeWidget_38 extends _i1.SmartFake implements _i9.Widget {
+  _FakeWidget_38(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 
   @override
@@ -281,18 +271,18 @@ class _FakeWidget_39 extends _i1.SmartFake implements _i9.Widget {
       super.toString();
 }
 
-class _FakeSentryOptions_40 extends _i1.SmartFake implements _i2.SentryOptions {
-  _FakeSentryOptions_40(Object parent, Invocation parentInvocation)
+class _FakeSentryOptions_39 extends _i1.SmartFake implements _i2.SentryOptions {
+  _FakeSentryOptions_39(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeScope_41 extends _i1.SmartFake implements _i2.Scope {
-  _FakeScope_41(Object parent, Invocation parentInvocation)
+class _FakeScope_40 extends _i1.SmartFake implements _i2.Scope {
+  _FakeScope_40(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
-class _FakeHub_42 extends _i1.SmartFake implements _i2.Hub {
-  _FakeHub_42(Object parent, Invocation parentInvocation)
+class _FakeHub_41 extends _i1.SmartFake implements _i2.Hub {
+  _FakeHub_41(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
@@ -1651,7 +1641,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i10.PipelineOwner get pipelineOwner => (super.noSuchMethod(
         Invocation.getter(#pipelineOwner),
-        returnValue: _FakePipelineOwner_26(
+        returnValue: _i14.dummyValue<_i10.PipelineOwner>(
           this,
           Invocation.getter(#pipelineOwner),
         ),
@@ -1660,7 +1650,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i10.RenderView get renderView => (super.noSuchMethod(
         Invocation.getter(#renderView),
-        returnValue: _FakeRenderView_27(
+        returnValue: _FakeRenderView_26(
           this,
           Invocation.getter(#renderView),
         ),
@@ -1669,7 +1659,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i10.MouseTracker get mouseTracker => (super.noSuchMethod(
         Invocation.getter(#mouseTracker),
-        returnValue: _FakeMouseTracker_28(
+        returnValue: _FakeMouseTracker_27(
           this,
           Invocation.getter(#mouseTracker),
         ),
@@ -1678,7 +1668,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i10.PipelineOwner get rootPipelineOwner => (super.noSuchMethod(
         Invocation.getter(#rootPipelineOwner),
-        returnValue: _FakePipelineOwner_26(
+        returnValue: _i14.dummyValue<_i10.PipelineOwner>(
           this,
           Invocation.getter(#rootPipelineOwner),
         ),
@@ -1699,7 +1689,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i9.PlatformMenuDelegate get platformMenuDelegate => (super.noSuchMethod(
         Invocation.getter(#platformMenuDelegate),
-        returnValue: _FakePlatformMenuDelegate_29(
+        returnValue: _FakePlatformMenuDelegate_28(
           this,
           Invocation.getter(#platformMenuDelegate),
         ),
@@ -1753,7 +1743,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i9.FocusManager get focusManager => (super.noSuchMethod(
         Invocation.getter(#focusManager),
-        returnValue: _FakeFocusManager_30(
+        returnValue: _FakeFocusManager_29(
           this,
           Invocation.getter(#focusManager),
         ),
@@ -2013,7 +2003,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
               ),
               (T v) => _i11.Future<T>.value(v),
             ) ??
-            _FakeFuture_31<T>(
+            _FakeFuture_30<T>(
               this,
               Invocation.method(
                 #scheduleTask,
@@ -2264,7 +2254,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           },
         ),
         returnValue: _i11.Future<_i6.Codec>.value(
-          _FakeCodec_32(
+          _FakeCodec_31(
             this,
             Invocation.method(
               #instantiateImageCodecFromBuffer,
@@ -2291,7 +2281,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
           {#getTargetSize: getTargetSize},
         ),
         returnValue: _i11.Future<_i6.Codec>.value(
-          _FakeCodec_32(
+          _FakeCodec_31(
             this,
             Invocation.method(
               #instantiateImageCodecWithSize,
@@ -2319,7 +2309,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i12.SemanticsHandle ensureSemantics() => (super.noSuchMethod(
         Invocation.method(#ensureSemantics, []),
-        returnValue: _FakeSemanticsHandle_33(
+        returnValue: _FakeSemanticsHandle_32(
           this,
           Invocation.method(#ensureSemantics, []),
         ),
@@ -2342,7 +2332,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   _i6.SemanticsUpdateBuilder createSemanticsUpdateBuilder() =>
       (super.noSuchMethod(
         Invocation.method(#createSemanticsUpdateBuilder, []),
-        returnValue: _FakeSemanticsUpdateBuilder_34(
+        returnValue: _FakeSemanticsUpdateBuilder_33(
           this,
           Invocation.method(#createSemanticsUpdateBuilder, []),
         ),
@@ -2351,7 +2341,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i10.PipelineOwner createRootPipelineOwner() => (super.noSuchMethod(
         Invocation.method(#createRootPipelineOwner, []),
-        returnValue: _FakePipelineOwner_26(
+        returnValue: _i14.dummyValue<_i10.PipelineOwner>(
           this,
           Invocation.method(#createRootPipelineOwner, []),
         ),
@@ -2375,7 +2365,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   ) =>
       (super.noSuchMethod(
         Invocation.method(#createViewConfigurationFor, [renderView]),
-        returnValue: _FakeViewConfiguration_35(
+        returnValue: _FakeViewConfiguration_34(
           this,
           Invocation.method(#createViewConfigurationFor, [renderView]),
         ),
@@ -2384,7 +2374,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i6.SceneBuilder createSceneBuilder() => (super.noSuchMethod(
         Invocation.method(#createSceneBuilder, []),
-        returnValue: _FakeSceneBuilder_36(
+        returnValue: _FakeSceneBuilder_35(
           this,
           Invocation.method(#createSceneBuilder, []),
         ),
@@ -2393,7 +2383,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i6.PictureRecorder createPictureRecorder() => (super.noSuchMethod(
         Invocation.method(#createPictureRecorder, []),
-        returnValue: _FakePictureRecorder_37(
+        returnValue: _FakePictureRecorder_36(
           this,
           Invocation.method(#createPictureRecorder, []),
         ),
@@ -2402,7 +2392,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i6.Canvas createCanvas(_i6.PictureRecorder? recorder) => (super.noSuchMethod(
         Invocation.method(#createCanvas, [recorder]),
-        returnValue: _FakeCanvas_38(
+        returnValue: _FakeCanvas_37(
           this,
           Invocation.method(#createCanvas, [recorder]),
         ),
@@ -2502,7 +2492,7 @@ class MockWidgetsFlutterBinding extends _i1.Mock
   @override
   _i9.Widget wrapWithDefaultView(_i9.Widget? rootWidget) => (super.noSuchMethod(
         Invocation.method(#wrapWithDefaultView, [rootWidget]),
-        returnValue: _FakeWidget_39(
+        returnValue: _FakeWidget_38(
           this,
           Invocation.method(#wrapWithDefaultView, [rootWidget]),
         ),
@@ -2575,7 +2565,7 @@ class MockHub extends _i1.Mock implements _i2.Hub {
   @override
   _i2.SentryOptions get options => (super.noSuchMethod(
         Invocation.getter(#options),
-        returnValue: _FakeSentryOptions_40(
+        returnValue: _FakeSentryOptions_39(
           this,
           Invocation.getter(#options),
         ),
@@ -2595,7 +2585,7 @@ class MockHub extends _i1.Mock implements _i2.Hub {
   @override
   _i2.Scope get scope => (super.noSuchMethod(
         Invocation.getter(#scope),
-        returnValue: _FakeScope_41(this, Invocation.getter(#scope)),
+        returnValue: _FakeScope_40(this, Invocation.getter(#scope)),
       ) as _i2.Scope);
 
   @override
@@ -2734,7 +2724,7 @@ class MockHub extends _i1.Mock implements _i2.Hub {
   @override
   _i2.Hub clone() => (super.noSuchMethod(
         Invocation.method(#clone, []),
-        returnValue: _FakeHub_42(this, Invocation.method(#clone, [])),
+        returnValue: _FakeHub_41(this, Invocation.method(#clone, [])),
       ) as _i2.Hub);
 
   @override