+++ flutter/test/profiling_test.dart
@@ -5,9 +5,9 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/profiling.dart';
+import 'package:sentry/src/platform/mock_platform.dart';
 import 'mocks.dart';
 import 'mocks.mocks.dart';
-import 'sentry_flutter_test.dart';
 
 void main() {
   late MockSentryNativeBinding mock;
@@ -20,7 +20,8 @@ void main() {
   group('$SentryNativeProfilerFactory', () {
     Hub hubWithSampleRate(double profilesSampleRate) {
       final o = defaultTestOptions();
-      o.platformChecker = getPlatformChecker(platform: MockPlatform.iOS());
+      o.platform = MockPlatform.iOS();
+      
       o.profilesSampleRate = profilesSampleRate;
 
       final hub = MockHub();